Develop the admin page of the widget corp.

+++ PHP-Rookie/databases_insert.php
@@ -17,8 +17,9 @@
 		 	//Often are form values in $_POST
 			$menu_name = "Today's Widget Trivia";
 			$position = (int) 4;
-			$visible = 1;
+			$visible = (int) 1;
 
+			//Escape all strings
 			$menu_name = mysqli_real_escape_string($connection,$menu_name);
 
 		 	//2. Perform database query
